+++ apps/pixelflut-cli/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "pixelflut-cli"
+
+authors = ["NotNite"]
+description = "A pixelflut client in Rust"
+edition = "2021"
+version = "0.1.0"
+
+[dependencies]
+clap = {version = "3.1.18", features = ["derive"]}
+image = "0.24.2"
+pixelflut = {path = "../../crates/pixelflut"}
